+++ src/Exception/InvalidQueryException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Scriptotek\Alma\Exception;
+
+class InvalidQueryException extends ClientException
+{
+}